pyproject.toml
[project]
name = "google-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.17.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.2",
]
